+++ pyVoIP/__init__.py
@@ -27,6 +27,7 @@ def debug(s, e=None):
     elif e is not None:
         print(f"[pyVoIP {stamp}] {e}")
 
+
 # noqa because import will fail if debug is not defined
 from pyVoIP.RTP import PayloadType  # noqa: E402